+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.12.0
  - feat: adding `NesBlinker`.
+ - fix: `NesRunningText` would not update its text.
 
 # 0.11.0
  - feat: add ability to customize `NesContainer`s pixel sizes.